getNativeContainerIds() now only returns entries with defined values.
In the original code, the return value of the getNativeContainerIds() can vary.
If sync was not enabled (or synced bookmarks we not loaded yet) it would return a map of entries that always have a defined value.
With the synced bookmarks present, it would returns entires with value=undefined for unsupported containers which did not currently have a mount-point bookmark folder (for example, due to the folder being deleted or renamed via the browser UI).
This logic was used in
  app/src/modules/webext/shared/webext-bookmark/webext-bookmark.service.ts
  Lines 1033 to 1044 in 317e697
method wasContainerChanged, which was called in processChangeType{Add,Modify.Remove}OnBookmarks methods
and also directly in processChangeTypeMoveOnBookmarks.
The code there checked whether the getNativeContainerIds had any undefined values, as a signal that a previously mapped natively unsupported container was somehow modified and not being detected by the getNativeContainerIds() any more.

Now, getNativeContainers() returns only entires for containers for which a native bookmark node id was discovered.

The code of wasContainerChanged mehod does the same check as before, but on its own - the logic (for now) remains equal to the old one!
It looks for any (synced)bookmarks.children (all containers in the current syncdata) that are not present in the result of getNativeContainerIds().
Also, wasContainerChanged() is now also used in processChangeTypeMoveOnBookmarks (unifying it with the other three processChange* calls)